actualizar usuario andando joya (faltan cositas)

+++ flaskps/models/usuario_tiene_rol.py
@@ -15,16 +15,18 @@ class User_tiene_rol(db.Model):
     
     def asignar_rol(idusuario, idrol):
         #agrega tupla a la tabla
-        elemento = User_tiene_rol (usuario_id = idusuario, rol_id = idrol)
-        db.session.add (elemento)
-        db.session.commit()   
+        if not User_tiene_rol.query.filter_by(usuario_id=idusuario, rol_id=idrol).first():
+            elemento = User_tiene_rol (usuario_id = idusuario, rol_id = idrol)
+            db.session.add (elemento)
+            db.session.commit() 
         return True
 
 
     def desasignar_rol(uid, rid):
         #borra tupla de la tabla
-        User_tiene_rol.query.filter_by(usuario_id=uid, rol_id=rid).delete()
-        db.session.commit()
+        if User_tiene_rol.query.filter_by(usuario_id=uid, rol_id=rid).first():
+            User_tiene_rol.query.filter_by(usuario_id=uid, rol_id=rid).delete()
+            db.session.commit()
         return True 
 
     # El siguiente método se usa con los id's, para ello cada clase tiene un método por el cual
